Move modal logic into modal component
